Modified how roll works
Added a new sub-component called Roll.vue that contains the data for rolling a random number,
passing up the picked string from the array using an $emit.

This sub-component makes it easier to have logic in a single place for the roll and the only
data needed for the main component is the picked string from the roll. Everything else should
belong only within the sub-component.